+++ R/weigh.R
@@ -40,12 +40,11 @@
 #' polymod |>
 #'   weigh("dayofweek", target = c(5, 2), groups = list(1:5, c(0, 6)))
 #'
-#' @importFrom data.table copy
 #' @export
 #' @autoglobal
 weigh <- function(survey, by, target = NULL, groups = NULL, ...) {
   check_if_contact_survey(survey)
-  survey <- copy(survey)
+  survey <- copy_survey(survey)
   participants <- survey$participants
 
   if (!by %in% colnames(participants)) {